requirements.txt
attrs==21.4.0
cayenne-mqtt==1.1.0
certifi==2021.10.8
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
clr-loader==0.2.4
colorama==0.4.4
cryptography==37.0.4
Flask==2.1.2
Flask-HTTPAuth==4.7.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.5.1
MarkupSafe==2.1.1
numpy
paho-mqtt==1.6.1
psutil==5.9.1
pycayennelpp==2.4.0
pycparser==2.21
pycurl
pygame==2.1.0
pyrsistent==0.18.1
pyserial==3.5
pythonnet==3.0.0.post1
requests==2.27.1
rfc3986==1.5.0
sniffio==1.3.0
tb-mqtt-client==1.2
urllib3==1.26.9
Werkzeug==2.1.2
